+++ Tests/Unit/Domain/Model/Event/SingleEventTest.php
@@ -1816,4 +1816,22 @@ final class SingleEventTest extends UnitTestCase
 
         self::assertSame($value, $this->subject->getSlug());
     }
+
+    /**
+     * @test
+     */
+    public function isRegistrationAllowedWithoutDateInitiallyReturnsFalse(): void
+    {
+        self::assertFalse($this->subject->isRegistrationAllowedWithoutDate());
+    }
+
+    /**
+     * @test
+     */
+    public function setRegistrationAllowedWithoutDateSetsRegistrationAllowedWithoutDate(): void
+    {
+        $this->subject->setRegistrationAllowedWithoutDate(true);
+
+        self::assertTrue($this->subject->isRegistrationAllowedWithoutDate());
+    }
 }