+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
+- Add a checkbox to allow event registration without a date (#4334)
 - Add an `Event.slug` property (#4331)
 - Add `Registration::belongsToUser()` (#4255)
 - Provide downloads in the "my registrations" view (#4249, #4250, #4251)